create homolog branch

+++ installl-node-exporter/install-node_exporter.sh
@@ -3,19 +3,40 @@
 # Diretório de instalação do Node Exporter
 bin_dir="/opt/node_exporter"
 
+arch=$(uname -m)
+case $arch in
+armv5*) arch="armv5" ;;
+armv6*) arch="armv6" ;;
+armv7*) arch="arm" ;;
+aarch64) arch="arm64" ;;
+x86) arch="386" ;;
+x86_64) arch="amd64" ;;
+i686) arch="386" ;;
+i386) arch="386" ;;
+*)
+    echo "Arquitetura não suportada"
+    exit 1
+    ;;
+esac
+
+# Solicitação da porta ao usuário
+printf "Porta a ser usada pelo Node Exporter (padrão: 9100): "
+read -r port
+port=${port:-9100}  # Define o valor padrão para 9100 caso o usuário não forneça nenhum valor
+
 # Baixa a última versão do Node Exporter e extrai a versão
 version=$(curl -s https://api.github.com/repos/prometheus/node_exporter/releases/latest | grep tag_name | cut -d '"' -f 4)
 node_exporter_version=$(echo "${version}" | cut -d 'v' -f 2)
 
 # Baixa o Node Exporter
-curl -LO "https://github.com/prometheus/node_exporter/releases/download/${version}/node_exporter-${node_exporter_version}.linux-amd64.tar.gz"
+curl -LO "https://github.com/prometheus/node_exporter/releases/download/${version}/node_exporter-${node_exporter_version}.linux-${arch}.tar.gz"
 
 # Descompacta o Node Exporter
-tar -xvf node_exporter-"${node_exporter_version}".linux-amd64.tar.gz
+tar -xvf node_exporter-"${node_exporter_version}.linux-${arch}.tar.gz"
 
 # Copia o binário do Node Exporter
 mkdir -p /opt/node_exporter
-cp node_exporter-"${node_exporter_version}".linux-amd64/node_exporter "${bin_dir}"
+cp node_exporter-"${node_exporter_version}.linux-${arch}/node_exporter" "${bin_dir}"
 
 # Cria o serviço do Node Exporter
 cat <<EOF >/etc/systemd/system/node_exporter.service
@@ -28,7 +49,7 @@ Wants=network-online.target
 User=node_exporter
 Group=node_exporter
 Restart=on-failure
-type=simple
+Type=simple
 RestartSec=3
 EnvironmentFile=${bin_dir}/node_exporter.conf
 ExecStart=${bin_dir}/node_exporter \$OPTIONS 
@@ -45,20 +66,25 @@ EOF
 chmod 664 /etc/systemd/system/node_exporter.service
 
 # Cria arquivo de configuração do Node Exporter
-cat <<"EOF" >"${bin_dir}/node_exporter.conf"
-OPTIONS="--web.listen-address=:9100"
+cat <<EOF >"${bin_dir}/node_exporter.conf"
+OPTIONS="--web.disable-exporter-metrics \
+--web.listen-address=:${port}"
 EOF
 
-# Cria o usuário e grupo do Node Exporter
-useradd --no-create-home --shell /bin/false node_exporter &&
-    chown -R node_exporter:node_exporter "${bin_dir}"
+# Cria o usuário e grupo do Node Exporter caso não exista
+if ! id -u node_exporter &>/dev/null; then
+    useradd --no-create-home --shell /bin/false node_exporter &&
+        chown -R node_exporter:node_exporter "${bin_dir}"
+fi
 
-# Cria link simbólico para o binário do Node Exporter
-ln -s "${bin_dir}/node_exporter" /usr/bin/node_exporter
+# Cria link simbólico para o binário do Node Exporter no diretório /usr/bin caso não exista
+if ! [ -e "/usr/bin/node_exporter" ]; then
+    ln -s "${bin_dir}/node_exporter" /usr/bin/node_exporter
+fi
 
 # Habilita e inicia o serviço do Node Exporter
 systemctl daemon-reload && systemctl enable --now node_exporter.service
 
 # Limpa arquivos temporários
-rm node_exporter-"${node_exporter_version}".linux-amd64.tar.gz
-rm -rf node_exporter-"${node_exporter_version}".linux-amd64
+rm node_exporter-"${node_exporter_version}.linux-${arch}.tar.gz"
+rm -rf node_exporter-"${node_exporter_version}.linux-${arch}"
